requirements.txt
asgiref==3.8.1
attrs==23.2.0
branca==0.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
Django==5.0.6
folium==0.16.0
h11==0.14.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
outcome==1.3.0.post0
pandas==2.2.2
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
selenium==4.20.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wsproto==1.2.0
xyzservices==2024.4.0
